Implementing result pattern on application layer

+++ VideoGameStore/Application/Customers/Command/Delete/DeleteCustomerCommandHandler.cs
@@ -1,9 +1,11 @@
 ﻿using VideoGameStore.Application.Interfaces;
+using VideoGameStore.Domain.common;
+using VideoGameStore.Domain.common.Errors;
 using VideoGameStore.Domain.Entities;
 
 namespace VideoGameStore.Application.Customers.Command.Delete
 {
-    public class DeleteCustomerCommandHandler : ICommandHandler<DeleteCustomerCommand>
+    public class DeleteCustomerCommandHandler : ICommandHandler<DeleteCustomerCommand, Result<Customer>>
     {
         private readonly IGenericRepository<Customer> _repository;
         private readonly IUnitOfWork _unitOfWork;
@@ -12,10 +14,16 @@ namespace VideoGameStore.Application.Customers.Command.Delete
             _repository = repository;
             _unitOfWork = unitOfWork;
         }
-        public async Task Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
+        public async Task<Result<Customer>> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
         {
-            await _repository.DeleteAsync(request.Id);
-            await _unitOfWork.CompleteAsync();
+            var customer =await _repository.GetByIdAsync(request.Id);
+            if (customer != null)
+            {
+                await _repository.DeleteAsync(request.Id);
+                await _unitOfWork.CompleteAsync();
+                return Result.Success(customer);
+            }
+            return Result.Failure<Customer>(CustomerErrors.FailedToFetchCustomer); 
             
 
         }
