+++ VideoGameStore/Application/Games/Command/Delete/DeleteGameCommandHandler.cs
@@ -1,9 +1,11 @@
 ﻿using VideoGameStore.Application.Interfaces;
+using VideoGameStore.Domain.common;
+using VideoGameStore.Domain.common.Errors;
 using VideoGameStore.Domain.Entities;
 
 namespace VideoGameStore.Application.Games.Command.Delete;
 
-public class DeleteGameCommandHandler : ICommandHandler<DeleteGameCommand, string>
+public class DeleteGameCommandHandler : ICommandHandler<DeleteGameCommand, Result<Game>>
 {
     private readonly IGenericRepository<Game> _gameRepository;
     private readonly IUnitOfWork _unitOfWork;
@@ -13,10 +15,15 @@ public class DeleteGameCommandHandler : ICommandHandler<DeleteGameCommand, strin
         _unitOfWork = unitOfWork;
     }
 
-    public async Task<string> Handle(DeleteGameCommand request, CancellationToken cancellationToken)
+    public async Task<Result<Game>> Handle(DeleteGameCommand request, CancellationToken cancellationToken)
     {
-       await _gameRepository.DeleteAsync(request.id);
-        await _unitOfWork.CompleteAsync();
-        return "Successful";
+        var game = await _gameRepository.GetByIdAsync(request.id);
+        if (game != null)
+        {
+            await _gameRepository.DeleteAsync(request.id);
+            await _unitOfWork.CompleteAsync();
+            return Result.Success(game);
+        }
+        return Result.Failure<Game>(GameErrors.GameNotFetched);
     }
 }
